Staging

+++ app/Http/Controllers/News/NewsController.php
@@ -128,7 +128,7 @@ class NewsController extends Controller implements NewsServices
         $categories = Category::whereNull('deleted_at')
         ->where('is_active','=','1')
         ->whereJsonContains('types','news')
-        ->get(); 
+        ->get();
 //        $tags = Tag::all();
         //        return response()->json($users);
 
@@ -152,18 +152,20 @@ class NewsController extends Controller implements NewsServices
     {
         $data = $request->input();
         $news_paginations = array();
-
+        $keyArr[]= null;
         try {
-            foreach ($data['keywords'] as $t) {
-                if (is_numeric($t)){
-                    $keyArr[] =  $t;
-                }
-                else{
-                    $newKeyword = Keywords::create(['keywords'=>$t,
-                                                'created_at' => now(),
-                                                'created_by' => Auth::user()->uuid,
-                                            ]);
-                    $keyArr[] = $newKeyword->id;
+            if ($data['keywords']!=null){
+                foreach ($data['keywords'] as $t) {
+                    if (is_numeric($t)){
+                        $keyArr[] =  $t;
+                    }
+                    else{
+                        $newKeyword = Keywords::create(['keywords'=>$t,
+                            'created_at' => now(),
+                            'created_by' => Auth::user()->uuid,
+                        ]);
+                        $keyArr[] = $newKeyword->id;
+                    }
                 }
             }
 
@@ -233,12 +235,17 @@ class NewsController extends Controller implements NewsServices
                 );
                 $log->save();
             }
-            foreach ($data['tags'] as $t) {
-                $news->tags()->attach($t, ['created_by' => auth()->id()]);
+
+            if ($request->has('tags')){
+                foreach ($data['tags'] as $t) {
+                    $news->tags()->attach($t, ['created_by' => auth()->id()]);
+                }
             }
 
-            foreach ($keyArr as $k) {
-                $news->keywords()->attach($k, ['created_by' => auth()->id()]);
+            if ($keyArr!=null){
+                foreach ($keyArr as $k) {
+                    $news->keywords()->attach($k, ['created_by' => auth()->id()]);
+                }
             }
 
             if (count($news_paginations)) {
@@ -283,7 +290,7 @@ class NewsController extends Controller implements NewsServices
         $categories = Category::whereNull('deleted_at')
         ->where('is_active','=','1')
         ->whereJsonContains('types','news')->get();;
-//        $tags = Tag::all(); 
+//        $tags = Tag::all();
         $keywords = Keywords::all();
         $contributors = $news->users;
         $users = User::with(['roles'])->get();
@@ -316,7 +323,7 @@ class NewsController extends Controller implements NewsServices
             'content' => $data['content'][0]
         ];
 
-        
+
 
         $i = 1;
         if (count($data['title']) > 1 && isset($data['title'][$id])) {
@@ -332,7 +339,7 @@ class NewsController extends Controller implements NewsServices
                 $i++;
             }
         }
-        
+
         if (count($data['title']) > 1) {
             foreach ($data['title'] as $key => $value) {
                 if (in_array($key, [0, $id])) {
@@ -355,18 +362,20 @@ class NewsController extends Controller implements NewsServices
         $date = $data['date'];
         $time = $data['time'];
         $margeDate = date('Y-m-d H:i:s', strtotime("$date $time"));
+        $keyArr[] = null;
         try {
-
-            foreach ($data['keywords'] as $t) {
-                if (is_numeric($t)){
-                    $keyArr[] =  $t;
-                }
-                else{
-                    $newKeyword = Keywords::create(['keywords'=>$t,
-                                                'created_at' => now(),
-                                                'created_by' => Auth::user()->uuid,
-                                            ]);
-                    $keyArr[] = $newKeyword->id;
+            if ($data['keywords']!=null){
+                foreach ($data['keywords'] as $t) {
+                    if (is_numeric($t)){
+                        $keyArr[] =  $t;
+                    }
+                    else{
+                        $newKeyword = Keywords::create(['keywords'=>$t,
+                            'created_at' => now(),
+                            'created_by' => Auth::user()->uuid,
+                        ]);
+                        $keyArr[] = $newKeyword->id;
+                    }
                 }
             }
 
@@ -410,13 +419,18 @@ class NewsController extends Controller implements NewsServices
                 $input['image'] = explode(Storage::url(""), $data['upload_image_selected'])[1];
             }
             $newsById->update($input);
-            $newsById::find($id)->tags()->detach();
-            foreach ($data['tags'] as $t) {
-                $newsById->tags()->attach($t, ['created_by' => auth()->id()]);
+            if ($request->has('tags')){
+                $newsById::find($id)->tags()->detach();
+                foreach ($data['tags'] as $t) {
+                    $newsById->tags()->attach($t, ['created_by' => auth()->id()]);
+                }
             }
-            $newsById::find($id)->keywords()->detach();
-            foreach ($keyArr as $k) {
-                $newsById->keywords()->attach($k, ['created_by' => auth()->id()]);
+
+            if ($keyArr!=null){
+                $newsById::find($id)->keywords()->detach();
+                foreach ($keyArr as $k) {
+                    $newsById->keywords()->attach($k, ['created_by' => auth()->id()]);
+                }
             }
 
             NewsPagination::upsert($news_paginations_old, ['id'], ['title', 'content', 'order_by_no']);
